chatbot/requirements.txt
flask
flask-cors
torch
git+https://github.com/openai/whisper.git
google-generativeai
python-dotenv 